Need to address at least two things here:

- [x] make faster for very small data sets as well (is this possible without dispatching to separate functions at some size threshold?)
- [x] add a unit test for the read-only buffer handling enhancement

~Much faster for large data sets, but slower for very small:~

```
       before           after         ratio
     [1f5dd630]       [28ed1501]
     <master>         <hausdorff_cython_improve>
+     29.6±0.08μs         61.3±3μs     2.07  spatial.Hausdorff.time_directed_hausdorff(10)
-       112±0.7μs         66.8±2μs     0.60  spatial.Hausdorff.time_directed_hausdorff(100)
-     1.32±0.01ms          466±1μs     0.35  spatial.Hausdorff.time_directed_hausdorff(1000)
```

The full summary of advantages, which I should also summarize in the final commit message if this is ever good enough:

- use `NPY_INFINITY` from a C header file to avoid overhead calls to `np.inf`
- add support for read-only array input
- remove the awkward `no_break_occurred` variable
- ~remove extraneous `i_store` and `j_store` variables~
- reduce 6 shuffling-related Python overhead calls to 2, by using `permutation` on the indices--there's no need to shuffle the arrays themselves as long as the iteration happens in shuffled order
- simplify the control flow guard on infinity propagation to `cmax` by setting `cmin=-inf` prior to a break

It would be amazing if we had access to true Cython-level shuffled range(). Perhaps the associated overhead will drop with new randomgen efforts upstream.