@tylerjereddy 2019-04-08
Faster across the board now:

```
       before           after         ratio
     [dadbbffd]       [8f91ce10]
     <master>         <hausdorff_cython_improve>
-        50.6±4μs         18.6±1μs     0.37  spatial.Hausdorff.time_directed_hausdorff(10)
-     2.18±0.09ms         550±10μs     0.25  spatial.Hausdorff.time_directed_hausdorff(1000)
-        183±10μs         32.0±2μs     0.17  spatial.Hausdorff.time_directed_hausdorff(100)
```

I'll place some comments / concerns inline.