@tylerjereddy 2019-04-09
Sounds like we don't have direct access to a faster Cython-level shuffler at the moment, and future improvements in that regard would have to wait until our minimum NumPy version catches up anyway.

If the CI remains all-green after my latest push, it would be helpful if @kain88-de could take a look just at the random-related changes, since he was driving the original constraints on what I did there.