@tylerjereddy 2019-05-21
I'll tentatively add 1.4.0 milestone for speed-up & functionality improvement. That said, I remain concerned about the random number adjustments I made here. Mixing the NumPy and C std library random machineries is probably not for the faint of heart.