@larsoner 2019-10-22
> Mixing the NumPy and C std library random machineries is probably not for the faint of heart.

Also, we'd want to make sure you get the same results/shuffling in SciPy 1.4.0 regardless of whether you have NumPy 1.16 or 1.17 installed (assuming this is the "before" and "after" for the randomgen refactoring they did). This sounds difficult.

So maybe just stick with one code path (C) for now?