@larsoner 2019-10-22
> So maybe just stick with one code path (C) for now?

... or I guess you could use the C code path if they pass an `int` as the seed, but allow them to pass a modern NumPy random generator instance if they have it. Such code will only work for users that have a sufficiently modern NumPy so no worries of version dependence.