@tylerjereddy 2019-11-12
I'm a little concerned about the complications here, and since it is just a performance improvement & my PR/not critical I'll just bump the milestone.