@tylerjereddy 2020-05-20
> I'm a little concerned about the complications here, and since it is just a performance improvement & my PR/not critical I'll just bump the milestone.

I'll make the final  decision that the benefits don't outweigh the downsides here--which is easy since I wrote the PR and will close it for good now :)